% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_graphs.R
\name{custom_bar_plot}
\alias{custom_bar_plot}
\title{My Custom Bar Plot}
\usage{
custom_bar_plot(data, x, fill, title = "Custom Bar Plot")
}
\arguments{
\item{data}{A data frame containing the data.}

\item{x}{Name of the variable for the x-axis (character).}

\item{fill}{Name of the variable for the fill (character).}

\item{title}{Title for the plot (default: "Custom Bar Plot").}
}
\value{
The output from \code{\link{print}}
}
\description{
This will create a bar plot with default styling and a minimal theme.
}
