% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_graphs.R
\name{custom_scatter_plot}
\alias{custom_scatter_plot}
\title{My Custom Scatter Plot}
\usage{
custom_scatter_plot(data, x, y, title = "Custom Scatter Plot")
}
\arguments{
\item{data}{A data frame containing the data.}

\item{x}{Name of the variable for the x-axis (character).}

\item{y}{Name of the variable for the y-axis (character).}

\item{title}{Title for the plot (default: "Custom Scatter Plot").}
}
\value{
The output from \code{\link{print}}
}
\description{
This will create a scatter plot with default orange points and a minimal theme.
}
